Xamarin EN android/ios の mock を作成

+++ Covid19Radar/Xamarin.ExposureNotification/ExposureNotification.ios.cs
@@ -247,7 +247,7 @@ namespace Xamarin.ExposureNotifications
 			var summary = new ExposureDetectionSummary(
 				(int)detectionSummary.DaysSinceLastExposure,
 				detectionSummary.MatchedKeyCount,
-				maxRisk,
+				(byte)maxRisk,
 				attDurTs.ToArray(),
 				sumRisk);
